% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{linkrepo}
\alias{linkrepo}
\title{Create link repo on Github}
\usage{
linkrepo(urls, repo)
}
\arguments{
\item{urls}{Vector of URLS}

\item{repo}{Name of GH repo.}
}
\description{
Creates a Github repository with a README full of links
}
